Extract operator expressions into own file
