+++ decoder/expr_any_semtok.go
@@ -9,8 +9,6 @@ import (
 	"github.com/hashicorp/hcl-lang/lang"
 	"github.com/hashicorp/hcl-lang/schema"
 	"github.com/hashicorp/hcl/v2/hclsyntax"
-	"github.com/zclconf/go-cty/cty"
-	"github.com/zclconf/go-cty/cty/convert"
 )
 
 func (a Any) SemanticTokens(ctx context.Context) []lang.SemanticToken {
@@ -138,56 +136,4 @@ func (a Any) semanticTokensForNonComplexExpr(ctx context.Context) []lang.Semanti
 		pathCtx: a.pathCtx,
 	}
 	return lt.SemanticTokens(ctx)
-}
-
-func (a Any) semanticTokensForOperatorExpr(ctx context.Context) ([]lang.SemanticToken, bool) {
-	tokens := make([]lang.SemanticToken, 0)
-
-	switch eType := a.expr.(type) {
-	case *hclsyntax.BinaryOpExpr:
-		opReturnType := eType.Op.Type
-
-		// Check if such an operation is even allowed within the constraint
-		if _, err := convert.Convert(cty.UnknownVal(opReturnType), a.cons.OfType); err != nil {
-			return tokens, true
-		}
-
-		opFuncParams := eType.Op.Impl.Params()
-		if len(opFuncParams) != 2 {
-			// This should never happen if HCL implementation is correct
-			return tokens, true
-		}
-
-		tokens = append(tokens, newExpression(a.pathCtx, eType.LHS, schema.AnyExpression{
-			OfType: opFuncParams[0].Type,
-		}).SemanticTokens(ctx)...)
-
-		tokens = append(tokens, newExpression(a.pathCtx, eType.RHS, schema.AnyExpression{
-			OfType: opFuncParams[1].Type,
-		}).SemanticTokens(ctx)...)
-
-		return tokens, true
-
-	case *hclsyntax.UnaryOpExpr:
-		opReturnType := eType.Op.Type
-
-		// Check if such an operation is even allowed within the constraint
-		if _, err := convert.Convert(cty.UnknownVal(opReturnType), a.cons.OfType); err != nil {
-			return tokens, true
-		}
-
-		opFuncParams := eType.Op.Impl.Params()
-		if len(opFuncParams) != 1 {
-			// This should never happen if HCL implementation is correct
-			return tokens, true
-		}
-
-		tokens = append(tokens, newExpression(a.pathCtx, eType.Val, schema.AnyExpression{
-			OfType: opFuncParams[0].Type,
-		}).SemanticTokens(ctx)...)
-
-		return tokens, true
-	}
-
-	return tokens, false
 }
